+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Features
 
 ## Fixes
+- Fixed npm_recording_extractor.py bug that was caused by mixing standard event CSV type CSVs with NPM data. [PR #256](https://github.com/LernerLab/GuPPy/pull/256)
 
 ## Deprecations and Removals